+++ demo/Calabonga.Wpf.MvvmMdi/Core/IVersionService.cs
@@ -14,7 +14,7 @@
     /// </summary>
     public class VersionService : IVersionService
     {
-        public VersionService() => Version = "1.0.7";
+        public VersionService() => Version = "1.0.0";
 
         public string Version { get; }
     }
