+++ demo/Calabonga.Wpf.MvvmMdi/Views/DetailZoneView.xaml
@@ -3,18 +3,36 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
+             d:DataContext="{d:DesignInstance viewModels:DetailsZoneViewModel}"
+             
+             xmlns:viewModels="clr-namespace:Calabonga.Wpf.MvvmMdi.ViewModels"
+             xmlns:b="http://schemas.microsoft.com/xaml/behaviors"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800" >
-    
-    <TextBox Text="Details Screen Content" Background="Coral" 
-             Foreground="Brown"
-             HorizontalContentAlignment="Center"
-             VerticalContentAlignment="Center">
-        <i:Interaction.Triggers>
-            <i:EventTrigger EventName="MouseDoubleClick">
-                <i:InvokeCommandAction Command="{Binding SendMessageCommand}"/>
-            </i:EventTrigger>
-        </i:Interaction.Triggers>
-    </TextBox>
+
+    <Grid HorizontalAlignment="Stretch" VerticalAlignment="Stretch" >
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+      
+        <TextBlock Text="DetailsZone"
+                   FontSize="20"
+                   TextAlignment="Center"
+                   Background="Brown" 
+                   Foreground="Coral"/>
+
+        <TextBox Text="DoubleClick to send Message to MainZone"
+                 Grid.Row="1"
+                 Background="Brown" 
+                 Foreground="Coral"
+                 HorizontalContentAlignment="Center"
+                 VerticalContentAlignment="Center">
+            <b:Interaction.Triggers>
+                <b:EventTrigger EventName="MouseDoubleClick">
+                    <b:InvokeCommandAction Command="{Binding SendMessageCommand}"/>
+                </b:EventTrigger>
+            </b:Interaction.Triggers>
+        </TextBox>
+    </Grid>
 </UserControl>
