+++ demo/Calabonga.Wpf.MvvmMdi/Views/MainZoneView.xaml
@@ -7,16 +7,33 @@
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
 
-    <TextBox Text="{Binding Message, TargetNullValue='Main Screen Content (Double Click to Close)'}"  Background="BlueViolet"
+    <Grid HorizontalAlignment="Stretch"
+          VerticalAlignment="Stretch">
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+
+        <TextBlock Text="MainZone"
+                   FontSize="20"
+                   TextAlignment="Center"
+                   Background="BlueViolet"
+                   Foreground="Aqua"/>
+
+        <TextBox Text="{Binding Message, TargetNullValue='DoubleClick to close MainZone module'}" 
+                 Background="BlueViolet"
                  Foreground="Aqua"
+                 Grid.Row="1"
                  HorizontalContentAlignment="Center"
                  VerticalContentAlignment="Center">
-        <b:Interaction.Triggers>
-            <b:EventTrigger EventName="MouseDoubleClick">
-                <b:InvokeCommandAction Command="{Binding CloseCommand}"/>
-            </b:EventTrigger>
-        </b:Interaction.Triggers>
-    </TextBox>
+            <b:Interaction.Triggers>
+                <b:EventTrigger EventName="MouseDoubleClick">
+                    <b:InvokeCommandAction Command="{Binding CloseCommand}"/>
+                </b:EventTrigger>
+            </b:Interaction.Triggers>
+        </TextBox>
+    </Grid>
 
 
 </UserControl>
