+++ src/Calabonga.Wpf.Mvvm.Zones/Calabonga.Wpf.Mvvm.Zones.csproj
@@ -4,7 +4,7 @@
 		<TargetFramework>net8.0-windows</TargetFramework>
 		<Nullable>enable</Nullable>
 		<UseWPF>true</UseWPF>
-		<Version>1.0.0-beta.2</Version>
+		<Version>1.0.0</Version>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
 		<Title>$(AssemblyName)</Title>
