+++ src/Calabonga.Wpf.Mvvm.Zones/MvvmObjectFactory.cs
@@ -12,6 +12,14 @@ public class MvvmObjectFactory : IMvvmObjectFactory
 
     public MvvmObjectFactory(IServiceProvider serviceProvider) => _serviceProvider = serviceProvider;
 
+    /// <summary>
+    /// Returns an instance of the View with ViewModel as DataContext assigned
+    /// </summary>
+    /// <typeparam name="TView"></typeparam>
+    /// <typeparam name="TViewModel"></typeparam>
+    /// <param name="onViewCreated"></param>
+    /// <param name="onViewModelCreated"></param>
+    /// <returns></returns>
     public IZoneView Create<TView, TViewModel>(Action<TView>? onViewCreated = null, Action<TViewModel>? onViewModelCreated = null)
         where TView : IZoneView
         where TViewModel : IZoneViewModel
@@ -29,6 +37,14 @@ public class MvvmObjectFactory : IMvvmObjectFactory
         return view;
     }
 
+    /// <summary>
+    /// Returns an instance of the ViewType with instance of the ViewModel as DataContext property.
+    /// </summary>
+    /// <param name="viewType"></param>
+    /// <param name="viewModelType"></param>
+    /// <param name="onViewCreated"></param>
+    /// <param name="onViewModelCreated"></param>
+    /// <returns></returns>
     public IZoneView Create(Type viewType, Type viewModelType, Action<object>? onViewCreated = null, Action<object>? onViewModelCreated = null)
     {
         using var scope = _serviceProvider.CreateScope();
