+++ src/Calabonga.Wpf.Mvvm.Zones/ServiceCollectionExtensions.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Calabonga.Wpf.Mvvm.Zones;
+
+public static class ServiceCollectionExtensions
+{
+    public static void AddZones(this IServiceCollection source) => source.AddSingleton<IZoneManager, ZoneManager>();
+}
